+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Bugs
 * Fix #5125: TLS 1.3 only should be supported
 * Fix #5145: [java-generator] handle `additionalProperties: true` emitting a field of type `AnyType`
+* Fix #5164: [java-generator] handle more special characters in field names
 
 #### Improvements